+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Added support to manually set bulk size and enable/disable bulk dispatch
 - `loadConfigurations` and `loadConfigurationsFromJson` functions
 
+### Fixed
+- Issue with plain log requests fixed
+
 ## [1.0.0-beta.1]
 ### Changes
  - Initial release